feat: Moved to skelton instead of bootstrap
I moved this project from using Boostrap to Skeleton becuase I felt that
the grid style and simplicity fit my needs better. I also decided to implement
ReactJS. This is of course a decision based on my desire for templating, React
seems to be the best implmentation I have come across so far.


bug: fixed inputs having wrong labels

Inputs did not work correctly on screen readers because many of the labels were named
incorrectly.